ui: BoardView skeleton replacing DebugDump (P6.5)
Status line + present/absent split; panels follow. Round-trip test
retargeted from DebugDump to BoardView.
